refactor: refactor return type of evaluate method
Refactor the return type of the Macro::evaluate and Variable::evaluate
methods to use the void return type. Now, we can pass a reference of
EvaluateResult to the evaluate method, and the method can set the result
of the evaluation in this object. This changes use the stack variable
instead of the heap variable to store the result of the evaluation.