fix(test): fix hyperscan error in the crs test
The test must be run in a separate thread to avoid the thread local
scratch space of hyperscan was not correctly clone from the main thread.
Because the other test cases may use the hyperscan scanner, and the
thread local scratch was initialized by the other test cases in the main
thread, so if we run this test in the main thread, the scratch space
will be not correctly initialized.

Fixes #13