fix(antlr4): fix SecRuleUpdateActionById
Fix the SecRuleUpdateActionById to correctly handle the `id`, `phase`
and `msg` metadata.
There has two limitations: it cannot be used to change the ID or phase
of a rule. Only the actions that can appear only once are overwritten.
The actions that are allowed to appear multiple times in a list, will be
appended to the end of the list.