fix(transaction): fix transform cache rehash error
When the transform cache is rehashed, the std::optional was not call
move construction of elements, because the move construction has no
noexcept specifier. Even we add noexcept specifier, if the string buffer
of element is short object, the reference to the buffer may be
invalidated during rehashing. This change use std::unique_ptr to instead
of std::optional to fix the issue.