feat(antlr4): support alias action (#100)
The alias action is a compile-time directive that allows users to create
shortcuts or alternative names for MATCHED_OPTREE and MATCHED_VPTREE
that have complex paths. This enhancement simplifies rule definitions
and improves readability by enabling the use of aliases instead of
lengthy variable paths.
The scope of an alias is limited to the rule in which it is defined and
all its chained rules. This design ensures that aliases do not interfere
with each other across different rules.